+++ tools/parsers/puddle/join_action.hpp
@@ -44,8 +44,8 @@ class JoinAction : public Action
         void setVerbose() { verbose = true; }
     private:
         std::string group;
-        int head;
-        int start, end;
+        size_t head;
+        size_t start, end;
         std::string ruleName;
         std::string type;
         bool verbose;
